+++ scripts/class/mastery/mastery-topple.js
@@ -1,4 +1,0 @@
-let targetToken = macroActivity.targets.first();
-if (targetToken) {
-    await targetToken.actor.toggleStatusEffect('prone', {active: true});
-}